DX: extract `conftest.py`
